+++ pages/market/[boardId]/payment/index.tsx
@@ -1,12 +1,9 @@
-import { useState } from 'react'
-import TryPaymentPage from '../../../../src/components/units/market/payment/TryPayment.container'
 import UseditemPaymentpage from '../../../../src/components/units/market/payment/useditemPayment.container'
 
 export default function PaymentPage() {
   return (
     <>
-      {/* <UseditemPaymentpage /> */}
-      <TryPaymentPage />
+      <UseditemPaymentpage />
     </>
   )
 }
