--- pages/review/index.tsx
@@ -1,5 +1,0 @@
-import CreateReviewPage from '../../src/components/units/market/review/createReview.container'
-
-export default function CreateReview() {
-  return <CreateReviewPage />
-}
