+++ src/components/units/market/main/Main.presenter.tsx
@@ -2,9 +2,12 @@
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
 import Slider from 'react-slick'
-
 import * as S from './Main.styles'
-export default function MainUI() {
+import { useMoveToPage } from '../../../commons/hooks/useMoveToPage'
+
+export default function MainUI(props) {
+  const { moveToPage } = useMoveToPage()
+
   const settings = {
     dots: true,
     infinite: true,
@@ -13,6 +16,7 @@ export default function MainUI() {
     slidesToScroll: 1,
     autoplay: true,
   }
+
   return (
     <>
       <S.Position>
@@ -27,44 +31,44 @@ export default function MainUI() {
               </S.Carousel>
             </Slider>
           </S.CarouselBox>
-          <S.NewArrivalsTxtBox>
-            <S.NewArrivalsTitleTxt>New Arrivals</S.NewArrivalsTitleTxt>
-            <S.NewArrivalsSubTxt>
-              새로 입고된 아이템들을 만나보세요!
-            </S.NewArrivalsSubTxt>
-          </S.NewArrivalsTxtBox>
+          <S.ResponsiveBox>
+            <S.NewArrivalsTxtBox>
+              <S.NewArrivalsTitleTxt>New Arrivals</S.NewArrivalsTitleTxt>
+              <S.NewArrivalsSubTxt>
+                새로 입고된 아이템들을 만나보세요!
+              </S.NewArrivalsSubTxt>
+            </S.NewArrivalsTxtBox>
 
-          <S.ProductBox>
-            {new Array(9).fill(0).map(() => (
-              <S.Product>
-                <S.ProductImg />
-                <S.ProductTxt>
-                  샤넬 유광 그레인드 카프스킨, 골드메탈 ..
-                </S.ProductTxt>
-                <S.ProductTxt>24,500,000원</S.ProductTxt>
-              </S.Product>
-            ))}
-          </S.ProductBox>
-          {/* <S.BannerChanelImg src="/images/MainPgChanelBannerImg.png" /> */}
-          <S.NewArrivalsTxtBox>
-            <S.NewArrivalsTitleTxt>BEST PRODUCT</S.NewArrivalsTitleTxt>
-            <S.NewArrivalsSubTxt>
-              지금 핫한 상품들을 만나보세요!
-            </S.NewArrivalsSubTxt>
-          </S.NewArrivalsTxtBox>
-          <S.ProductBox>
-            {new Array(9).fill(0).map(() => (
-              <S.Product>
-                <S.ProductImg />
-                <S.ProductTxt>
-                  샤넬 유광 그레인드 카프스킨, 골드메탈 ..
-                </S.ProductTxt>
-                <S.ProductTxt>24,500,000원</S.ProductTxt>
-              </S.Product>
-            ))}
-          </S.ProductBox>
+            <S.ProductBox>
+              {props.NewArrival?.map((el) => (
+                <S.Product onClick={moveToPage(`market/${el.id}`)}>
+                  <S.ProductImg />
+                  <S.ProductTxt>{el.name}</S.ProductTxt>
+                  <S.ProductTxt>{el.price}원</S.ProductTxt>
+                </S.Product>
+              ))}
+            </S.ProductBox>
+            {/* <S.BannerChanelImg src="/images/MainPgChanelBannerImg.png" /> */}
+            <S.NewArrivalsTxtBox>
+              <S.NewArrivalsTitleTxt>BEST PRODUCT</S.NewArrivalsTitleTxt>
+              <S.NewArrivalsSubTxt>
+                지금 핫한 상품들을 만나보세요!
+              </S.NewArrivalsSubTxt>
+            </S.NewArrivalsTxtBox>
+            <S.ProductBox>
+              {new Array(9).fill(0).map(() => (
+                <S.Product>
+                  <S.ProductImg />
+                  <S.ProductTxt>
+                    샤넬 유광 그레인드 카프스킨, 골드메탈 ..
+                  </S.ProductTxt>
+                  <S.ProductTxt>24,500,000원</S.ProductTxt>
+                </S.Product>
+              ))}
+            </S.ProductBox>
 
-          <S.LoadMoreBtn>더보기</S.LoadMoreBtn>
+            <S.LoadMoreBtn>더보기</S.LoadMoreBtn>
+          </S.ResponsiveBox>
         </S.Wrapper>
       </S.Position>
     </>
