+++ src/components/units/market/main/Main.queries.ts
@@ -1,0 +1,13 @@
+import { gql } from '@apollo/client'
+
+export const FETCH_PRODUCTS = gql`
+  query fetchProducts {
+    fetchProducts {
+      id
+      name
+      description
+      price
+      urls
+    }
+  }
+`
